UX: Network utilization color
Resolves #2926
