+++ lib/networks/getNetworkUtilizationParams.ts
@@ -1,0 +1,25 @@
+export default function getNetworkUtilizationParams(value: number) {
+  const load = (() => {
+    if (value > 80) {
+      return 'high';
+    }
+
+    if (value > 50) {
+      return 'medium';
+    }
+
+    return 'low';
+  })();
+
+  const colors = {
+    high: 'red.600',
+    medium: 'orange.600',
+    low: 'green.600',
+  };
+  const color = colors[load];
+
+  return {
+    load,
+    color,
+  };
+}
