+++ ui/gasTracker/GasTrackerNetworkUtilization.tsx
@@ -1,6 +1,7 @@
 import { chakra } from '@chakra-ui/react';
 import React from 'react';
 
+import getNetworkUtilizationParams from 'lib/networks/getNetworkUtilizationParams';
 import { Skeleton } from 'toolkit/chakra/skeleton';
 import { mdash } from 'toolkit/utils/htmlEntities';
 
@@ -10,24 +11,7 @@ interface Props {
 }
 
 const GasTrackerNetworkUtilization = ({ percentage, isLoading }: Props) => {
-  const load = (() => {
-    if (percentage > 80) {
-      return 'high';
-    }
-
-    if (percentage > 50) {
-      return 'medium';
-    }
-
-    return 'low';
-  })();
-
-  const colors = {
-    high: 'red.600',
-    medium: 'orange.600',
-    low: 'green.600',
-  };
-  const color = colors[load];
+  const { load, color } = getNetworkUtilizationParams(percentage);
 
   return (
     <Skeleton loading={ isLoading } whiteSpace="pre-wrap">
